src/config.py
import os

from dotenv import load_dotenv

load_dotenv()

KNOWLEDGE_BASE_PATH = os.getenv("KNOWLEDGE_BASE_PATH")
EMBEDDING_MODEL_NAME = os.getenv("EMBEDDING_MODEL_NAME")
CHROMA_COLLECTION_NAME = os.getenv("CHROMA_COLLECTION_NAME")
CHROMA_HOST = os.getenv("CHROMA_HOST")
CHROMA_PORT = os.getenv("CHROMA_PORT")
