requirements.txt
aiogram>=3.0.0
aiogram_i18n>=2.0.0
alembic>=1.12.0
asyncpg>=0.28.0
redis>=5.0.0
sqlalchemy>=2.0.0
msgspec>=0.18.0
pydantic>=2.0.0
pydantic_settings>=2.0.0
fluent_runtime>=0.4.0
types-pyyaml>=6.0.0
tzdata>=2023.3
ipykernel>=6.0.0
ruff-lsp>=0.0.45
ruff>=0.1.0
black>=23.0.0
libcst>=1.0.0
openai>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
orjson>=3.9.0
python-dotenv>=1.0.0
uvloop>=0.19.0
aiofiles>=23.2.1
pymupdf4llm>=0.1.2
watchdog>=3.0.0

